feat: add animation and redesign the award-impact- partner
